Fix #663, Scrub datatypes across TIME
Fix all use of uint32 for sizes, replace with size_t
Scrub all other types to make sure use is consistent,
using proper typedef whenever available.

+++ fsw/cfe-core/src/time/cfe_time_task.c
@@ -427,10 +427,10 @@ int32 CFE_TIME_TaskInit(void)
 **  Return:
 **    true if length is acceptable
 */
-bool CFE_TIME_VerifyCmdLength(CFE_SB_MsgPtr_t Msg, uint16 ExpectedLength)
+bool CFE_TIME_VerifyCmdLength(CFE_SB_MsgPtr_t Msg, size_t ExpectedLength)
 {
     bool    result       = true;
-    uint16  ActualLength = CFE_SB_GetTotalMsgLength(Msg);
+    size_t  ActualLength = CFE_SB_GetTotalMsgLength(Msg);
 
     /*
     ** Verify the command packet length
